+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog: recurtools
 
+- removed lenrecursive function in preference for `len(nested())`. `countcontainers` not implemented in new version.
 - **rename** `chainanything` to `starchain`, **default behaviour changed** to `recursive = True`
 - standardised on keywork argument `preserve`